cli/login: move login_with_chatgpt to cli crate; print local login server URL
Relocated the async `login_with_chatgpt` function from codex-login into the CLI crate, updating it to print the browser auth URL and port when starting the local login server. Removed the lib.rs version. Uses mpsc to notify the user of the URL.
